Renamed tenants_only to database_tenants
