+++ pgrest/views.py
@@ -10,7 +10,7 @@ from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseServer
 import requests
 from rest_framework.views import APIView
 
-from tenants_only.models import Tenants
+from database_tenants.models import Tenants
 from pgrest.models import ManageTables, ManageTablesTransition, ManageViews
 from pgrest.db_transactions import manage_tables, table_data, bulk_data, view_data
 from pgrest.pycommon.auth import t, get_tenant_id_from_base_url
@@ -1070,35 +1070,6 @@ class DynamicViewById(RoleSessionMixin, APIView):
 
         return HttpResponse(make_success(msg="Row delected successfully."), content_type='application/json')
 
-class CreateTenant(APIView):
-    def post(self, request, *args, **kwargs):
-        logger.debug("top of post /manage/tenants")
-        try:
-            schema_name = request.data['schema_name']
-            db_instance = request.data['db_instance']
-        except KeyError as e:
-            msg = f"\'{e.args}\' is required when creating new row in a table."
-            logger.warning(msg)
-            return HttpResponseBadRequest(make_error(msg=msg))
-
-        try:
-            tenant = Tenants(schema_name = schema_name,
-                             tenant_name = schema_name,
-                             db_instance_name = db_instance)
-        except Exception as e:
-            msg = f"Unable to create new tenant for Tenants table"
-            logger.warning(msg + f" e: {e}")
-            return HttpResponseBadRequest(make_error(msg=msg))
-
-        try:
-            tenant.save()
-        except Exception as e:
-            msg = f"Failed to create new tenant {schema_name} in db_instance {db_instance}. {e}"
-            logger.error(msg)
-            return HttpResponseBadRequest(make_error(msg=msg))
-
-        return HttpResponse(make_success(msg=f"Tenant '{schema_name}' created successfully."), content_type='application/json')
-
 
 ### Views
 class ViewManagement(RoleSessionMixin, APIView):
@@ -1418,3 +1389,4 @@ class ViewsResource(RoleSessionMixin, APIView):
             return HttpResponseBadRequest(make_error(msg=msg))
 
         return HttpResponse(make_success(result=result), content_type='application/json')
+
